size

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,6 @@
-#include <stdio.h>
+#include<stdio.h>
 /**
- * main - A  c program that prints size of various computer types
+ * main - A c program that prints size of various computer types
  * return: 0 (Success)
  */
 int main(void)
@@ -11,10 +11,11 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of char: %lu byte(s)\n", sizeof(a));
+printf("Size of an int: %lu byte(s)\n", sizeof(b));
+printf("Size of a long int: %lu byte(s)\n", sizeof(c));
+printf("Size of a long long int: %lu byte(s)\n", sizeof(d));
+printf("Size of a float: %lu byte(s)\n", sizeof(f));
+
 return (0);
 }
